Added vehicle functions to Landrovers
