--- addons/jackal_scripts/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\LSR\addons\jackal_scripts
