--- addons/jackal_scripts/functions/fnc_vehicleInit.sqf
@@ -1,17 +1,0 @@
-/*
-Author: www.3commandobrigade.com
-1. Add CBA key bind for vehicle weapon reload.
-2. Add CBA key bind for high/low beam lights
-Arguments:
-none
-Return Value:
-none
-*/
-#include "script_component.hpp"
-["16aa", "16AA_Key_Reload_Vehicle_Weapon", localize LSTRING(ReloadWeapon),
-{
-     [ACE_player, vehicle ACE_player] call FUNC(resupplyAmmo);
-     vehicle ACE_player;
-},
-{false},
-[19, [false, true, false]]] call cba_fnc_addKeybind;
