--- addons/jackal_scripts/script_component.hpp
@@ -1,12 +1,0 @@
-#define COMPONENT jackal_scripts
-#include "\z\LSR\addons\main\script_mod.hpp"
-
-#ifdef DEBUG_ENABLED_JACKAL_SCRIPTS
-    #define DEBUG_MODE_FULL
-#endif
-
-#ifdef DEBUG_SETTINGS_JACKAL_SCRIPTS
-    #define DEBUG_SETTINGS DEBUG_SETTINGS_JACKAL_SCRIPTS
-#endif
-
-#include "\z\LSR\addons\main\script_macros.hpp"
