+++ addons/vehicle_scripts/$PBOPREFIX$
@@ -1,0 +1,1 @@
+z\LSR\addons\vehicle_scripts
