+++ addons/vehicle_scripts/XEH_preInit.sqf
