+++ addons/vehicle_scripts/functions/fnc_addMagazine.sqf
@@ -29,7 +29,7 @@ None
 	} forEach (_listOfMags select 0);
 
 
-	["16aa_jackals_addMagazine", [_vehicle,_magName ,_turretIndex]] call ace_common_fnc_globalEvent;
+	["16aa_vehicles_addMagazine", [_vehicle,_magName ,_turretIndex]] call ace_common_fnc_globalEvent;
