+++ addons/vehicle_scripts/functions/fnc_debug.sqf
