+++ addons/vehicle_scripts/functions/fnc_dismountGPMG.sqf
@@ -6,6 +6,7 @@ Removes the commanders GPMG from a jackal and adds it to the player
 Arguments:
 0: vehicle <OBJECT>
 1: unit <OBJECT>
+2: textureIndex <NUMBER>
 
 Return Value:
 Nothing
@@ -15,7 +16,7 @@ None
 */
 #include "script_component.hpp"
 
-params["_vehicle","_unit"];
+params["_vehicle","_unit","_textureIndex"];
 
 if !([_vehicle, _unit] call FUNC(canDismount)) exitwith {};
 
@@ -26,15 +27,15 @@ if ((_unit call CBA_fnc_getUnitAnim) select 0 == "stand") then {
 };
 
 [{
-    params["_vehicle","_unit"];
+    params["_vehicle","_unit","_textureIndex"];
 
-    ["16aa_jackals_setWeaponTurretGPMG", [_vehicle, false]] call ace_common_fnc_globalEvent;
+    ["16aa_vehicles_setWeaponTurretGPMG", [_vehicle, false]] call ace_common_fnc_globalEvent;
 
     _pos = _unit modelToWorldVisual [0,1,0.05];
     _unit = createVehicle ["WeaponHolder_Single_F",_pos,[],0,"NONE"];
     _unit addWeaponCargoGlobal ['16aa_l7a2',1];
     _unit setPosATL _pos;
 
-	_vehicle setObjectTextureGlobal [3, ""];
+	_vehicle setObjectTextureGlobal [_textureIndex, ""];
 
-}, [_vehicle, _unit], 1, 0]call ace_common_fnc_waitAndExecute;
+}, [_vehicle, _unit,_textureIndex], 1, 0]call ace_common_fnc_waitAndExecute;
