+++ addons/vehicle_scripts/functions/fnc_dismountGPMGTimer.sqf
@@ -1,10 +1,11 @@
 /*
 Author: Grey
-Mounts GPMG using a timer
+Dismount GPMG using a timer
 
 Arguments:
 0: vehicle <OBJECT>
 1: unit <OBJECT>
+2: textureIndex <NUMBER>
 
 Return Value:
 Nothing
@@ -14,10 +15,10 @@ Return value:
 */
 #include "script_component.hpp"
 
-params["_vehicle","_unit"];
+params["_vehicle","_unit","_textureIndex"];
 
 if ((_unit call CBA_fnc_getUnitAnim) select 0 == "stand") then {
     [_unit, "AmovPercMstpSrasWrflDnon_diary", 1] call ace_common_fnc_doAnimation;
 };
 
-[10, [_vehicle, _unit], {(_this select 0) call FUNC(mountGPMG);}, {}, "Mounting GPMG"] call ace_common_fnc_progressBar;
+[10, [_vehicle, _unit,_textureIndex], {(_this select 0) call FUNC(dismountGPMG);}, {}, "Dismounting GPMG"] call ace_common_fnc_progressBar;
