+++ addons/vehicle_scripts/functions/fnc_mountGPMG.sqf
@@ -6,6 +6,7 @@ Removes the a GPMG from players inventory and "mounts" it to a Jackal
 Arguments:
 0: vehicle <OBJECT>
 1: unit <OBJECT>
+2: textureIndex <NUMBER>
 
 Return Value:
 Nothing
@@ -15,7 +16,7 @@ None
 */
 #include "script_component.hpp"
 
-params["_vehicle","_unit"];
+params["_vehicle","_unit","_textureIndex"];
 
 if !([_vehicle, _unit] call FUNC(canDismount)) exitwith {};
 
@@ -30,8 +31,8 @@ if ((_unit call CBA_fnc_getUnitAnim) select 0 == "stand") then {
 
     _unit removeWeapon '16aa_l7a2';
 
-    ["16aa_jackals_setWeaponTurretGPMG", [_vehicle, true]] call ace_common_fnc_globalEvent;
+    ["16aa_vehicles_setWeaponTurretGPMG", [_vehicle, true]] call ace_common_fnc_globalEvent;
 
-	_vehicle setObjectTextureGlobal [3, "16aa_vehicles_common\weapons\gpmg\m240_co.paa"];
+	_vehicle setObjectTextureGlobal [_textureIndex, "16aa_vehicles_common\weapons\gpmg\m240_co.paa"];
 
-}, [_vehicle, _unit], 1, 0]call ace_common_fnc_waitAndExecute;
+}, [_vehicle, _unit,_textureIndex], 1, 0]call ace_common_fnc_waitAndExecute;
