+++ addons/vehicle_scripts/functions/script_component.hpp
