+++ addons/vehicle_scripts/script_component.hpp
@@ -1,0 +1,12 @@
+#define COMPONENT vehicle_scripts
+#include "\z\LSR\addons\main\script_mod.hpp"
+
+#ifdef DEBUG_ENABLED_VEHICLE_SCRIPTS
+    #define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_VEHICLE_SCRIPTS
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_VEHICLE_SCRIPTS
+#endif
+
+#include "\z\LSR\addons\main\script_macros.hpp"
