+++ addons/vehicle_scripts/stringtable.xml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project name="LSR">
-    <Package name="jackal_scripts">
-        <Key ID="STR_LSR_jackal_scripts_ReloadWeapon">
+    <Package name="vehicle_scripts">
+        <Key ID="STR_LSR_vehicle_scripts_ReloadWeapon">
             <English>Reload Vehicle Weapon</English>
         </Key>
     </Package>
